refactor: move out log file configuration
